Added comment about strictness of regex check
Someone asked me a question that made me think - what if the file input was actually YZX or ZXY instead of YXZ? Our regex checker probably would accept it, but it wouldn't be a valid YXZ file. We need to make it so we actually put in the valid range of latitude, longitude, and elevation values.